Split interface and class
